+++ index.css
@@ -1,200 +1,200 @@
 * {
-    box-sizing: border-box;
+  box-sizing: border-box;
 }
 
 body {
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    margin: 0.5rem;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  margin: 0.5rem;
 }
 
 header {
-    border: 0.3rem solid black;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    margin: 0;
-    justify-content: center;
-    align-items: center;
+  border: 0.3rem solid black;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  margin: 0;
+  justify-content: center;
+  align-items: center;
 }
 
 header h2 {
-    margin: 1rem 0 0 0;
+  margin: 1rem 0 0 0;
 }
 
 header ul {
-    display: inline-flex;
-    list-style: none;
-    gap: 1rem;
-    padding: 0;
+  display: inline-flex;
+  list-style: none;
+  gap: 1rem;
+  padding: 0;
 }
 
 #listBtn,
 #addBookBtn,
 #contactBtn {
-    text-decoration: none;
-    color: black;
+  text-decoration: none;
+  color: black;
 }
 
 #listBtn:hover,
 #addBookBtn:hover,
 #contactBtn:hover {
-    color: grey;
+  color: grey;
 }
 
 #list {
-    flex-direction: column;
-    margin: 0;
-    justify-content: center;
-    align-items: center;
+  flex-direction: column;
+  margin: 0;
+  justify-content: center;
+  align-items: center;
 }
 
 #list h2 {
-    text-align: center;
+  text-align: center;
 }
 
 #library {
-    border: 0.3rem solid black;
-    width: 100%;
+  border: 0.3rem solid black;
+  width: 100%;
 }
 
 .book {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
 }
 
 .book:nth-child(odd) {
-    background-color: lightgray;
+  background-color: lightgray;
 }
 
 .bookDetail {
-    display: inline-flex;
-    gap: 0.5rem;
-    margin-left: 1rem;
-    flex-wrap: wrap;
+  display: inline-flex;
+  gap: 0.5rem;
+  margin-left: 1rem;
+  flex-wrap: wrap;
 }
 
 .removeBtn {
-    margin: 0.1rem;
+  margin: 0.1rem;
 }
 
 .removeBtn:hover {
-    background-color: darkred;
-    color: white;
+  background-color: darkred;
+  color: white;
 }
 
 #addBook {
-    flex-direction: column;
-    padding: 5% 0;
-    align-items: center;
+  flex-direction: column;
+  padding: 5% 0;
+  align-items: center;
 }
 
 #addBook h2 {
-    text-align: center;
+  text-align: center;
 }
 
 form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 60%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 60%;
 }
 
 #title,
 #author {
-    width: 100%;
-    font-size: 0.8rem;
-    padding: 0.5rem;
-    border: 0.3rem solid black;
+  width: 100%;
+  font-size: 0.8rem;
+  padding: 0.5rem;
+  border: 0.3rem solid black;
 }
 
 #addBtn {
-    align-self: self-end;
+  align-self: self-end;
 }
 
 #addBtn:hover {
-    background-color: darkred;
-    color: white;
+  background-color: darkred;
+  color: white;
 }
 
 #contact {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 5% 15%;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 5% 15%;
 }
 
 #contact ul {
-    direction: ltr;
+  direction: ltr;
 }
 
 footer {
-    position: fixed;
-    bottom: 0;
-    border: 0.3rem solid black;
-    display: inline-flex;
-    margin: 0 0 1rem 0;
-    width: 98%;
-    z-index: 1;
-    background-color: white;
+  position: fixed;
+  bottom: 0;
+  border: 0.3rem solid black;
+  display: inline-flex;
+  margin: 0 0 1rem 0;
+  width: 98%;
+  z-index: 1;
+  background-color: white;
 }
 
 footer p {
-    margin-left: 2rem;
+  margin-left: 2rem;
 }
 
 @media screen and (min-width: 768px) {
-    header {
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-    }
+  header {
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  }
 
-    header h2 {
-        margin: 0 0 0 2rem;
-        font-size: 2rem;
-    }
+  header h2 {
+    margin: 0 0 0 2rem;
+    font-size: 2rem;
+  }
 
-    header ul {
-        margin-right: 2rem;
-        font-size: 1.5rem;
-    }
+  header ul {
+    margin-right: 2rem;
+    font-size: 1.5rem;
+  }
 
-    #list h2 {
-        font-size: 3rem;
-    }
+  #list h2 {
+    font-size: 3rem;
+  }
 
-    #library {
-        width: 80%;
-    }
+  #library {
+    width: 80%;
+  }
 
-    .book {
-        font-size: 1.5rem;
-    }
+  .book {
+    font-size: 1.5rem;
+  }
 
-    .removeBtn {
-        margin: 0.2rem;
-        font-size: 1.3rem;
-    }
+  .removeBtn {
+    margin: 0.2rem;
+    font-size: 1.3rem;
+  }
 
-    #addBook h2 {
-        font-size: 3rem;
-    }
+  #addBook h2 {
+    font-size: 3rem;
+  }
 
-    #title,
-    #author {
-        font-size: 2rem;
-    }
+  #title,
+  #author {
+    font-size: 2rem;
+  }
 
-    #addBtn {
-        font-size: 2rem;
-        padding: 1rem;
-    }
+  #addBtn {
+    font-size: 2rem;
+    padding: 1rem;
+  }
 
-    #contact {
-        font-size: 2rem;
-    }
+  #contact {
+    font-size: 2rem;
+  }
 }
