+++ index.html
@@ -51,7 +51,7 @@
     <footer>
         <p>&COPY; Copyright 2023 Awesome Books</p>
     </footer>
-    <script src="./index.js"></script>
+    <script src="./index.js" type="module"></script>
 </body>
 
 </html>
